Cleanup Dependencies, Add Requestor and Seeded Randomizer
The Requestor simplifies us pulling the name off a request, and then
reducing the name and date to a valid seed that will allow us to
reproduce the same results across requests for the same day. This
actually makes our randomized element of DAFTE deterministic, but it's
probably random enough for practical use, and allows us to have
idempotent daily advice.