Massive cleanup!
Removed the QueryParamsUtil, created a new DafteRequest to act as an
abstraction layer for all our supported request types. This allowed a
lot of cleanup when trying to pull params off the requests and removed
the need to overload the factories with both Request types.

Also, fixed the formatting for the date of the request.